@@ -1,55 +1,67 @@
 package com.example.euapp16.myapplication;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.AttributeSet;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+import android.support.v7.app.AppCompatActivity;
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.toolbox.JsonObjectRequest;
+import com.android.volley.toolbox.Volley;
+
+import org.json.JSONException;
+import org.json.JSONObject;
 
 public class MainActivity extends AppCompatActivity {
-
-    Button buttonLogin, buttonRegister;
-    EditText editTextName, editTextPasswd;
-    public Intent intentMain, intentSecond;
+    private RequestQueue loginqueue;
+    private EditText editTextName, editTextPasswd;
+    private Intent intentMain, intentSecond;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        loginqueue = Volley.newRequestQueue(this);
+
         /* Finds all required objects in the current View */
-        buttonLogin = findViewById(R.id.buttonLogin);
-        buttonRegister = findViewById(R.id.buttonRegister);
+        Button buttonLogin = findViewById(R.id.buttonLogin);
+        Button buttonRegister = findViewById(R.id.buttonRegister);
         editTextName = findViewById(R.id.editText);
         editTextPasswd = findViewById(R.id.editText2);
 
         /* Sends you off to the Registration activity */
 
-        buttonRegister.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                intentMain = new Intent (getApplicationContext(), logon_register.class);
-                startActivity(intentMain);
-            }
+        buttonRegister.setOnClickListener(v -> {
+            intentMain = new Intent(getApplicationContext(), logon_register.class);
+            startActivity(intentMain);
         });
-
         /* TODO - replace with database login check */
-        buttonLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {String n = editTextName.getText().toString();
-                String ph = editTextPasswd.getText().toString();
-                String x = "root";
-                if (n.equals(x) && ph.equals(x)) {
-                    intentSecond = new Intent (getApplicationContext(), Subject_activity.class);
-                    startActivity(intentSecond);
-                }
-                else {
-                    Toast.makeText(MainActivity.this, "E-mail or password wrong.", Toast.LENGTH_SHORT).show();
-                }
-            }
+        buttonLogin.setOnClickListener(v -> {
+            Intent intent = new Intent(MainActivity.this,Subject_activity.class);
+            loginParse(editTextName.getText().toString(), editTextPasswd.getText().toString(), intent);
         });
+    }
+
+    private void loginParse(String uname, String passwd, Intent intent){
+        String url = getString(R.string.server_address) + "/eu/YellowTeam/API/login.php?username=" + uname + "&password="+passwd;
+        JsonObjectRequest request = new JsonObjectRequest(Request.Method.POST, url, null,
+                response -> {
+                    try {
+                        JSONObject messageFromServer = response.getJSONObject("result");
+                        if (messageFromServer.getBoolean("enableLogin")&& uname.equals(messageFromServer.getString("username"))){
+                            startActivity(intent);
+                        }
+                        else{
+                            Toast.makeText(MainActivity.this,"Wrong username or password!",Toast.LENGTH_LONG).show();
+                        }
+
+                    } catch (JSONException e) {
+                        e.printStackTrace();
+                    }
+                }, Throwable::printStackTrace);
+        loginqueue.add(request);
     }
 }
